Module ssmp::api

Structs

ClientRequestAPI
Defines Client Request API. This will handle all requests sent to our Rocket server.

MultiQueryRecord
Struct that stores logged in user information Different from QueryResult since it's used when returning info for all users

QueryResult
Struct that we deserialize from JSON responses that contains user information for a given query.

Enums

APIResponse
General response structure, used in Enum form for flexibility.

Functions

mock_get_online_users
supply with a vector of public keys for users. always start with at least user jfdenton, port: 9090 will create users starting at port 9091 --> len(keys)

nat_punch_init
user the server for a NAT punch through.