Struct ssmp::messages::packet::Packet

pub struct Packet {
    pub header: PacketHeader,
    // some fields omitted
}

The SSMP Packet

The SSMP Packet is made up of three main sections:

The main header

The user header

The payload

Both the user-defined header and the payload are optional and mutually exclusive. All packets contain the main header, but only some packets have a user header and/or a payload.

If all three are present, the user header follows the main header, and the payload follows the user header. If there's no user header, the payload is directly after the main header.

The payload MUST be the last section of the packet, since the main header doesn't explicitly define a 'payload length' field.

A SSMP packet with a payload and user header would look like this:

//       +----------------+----------------+----------------+----------------+
// Byte  |       0        |       1        |       2        |       3        |
//       +================+================+================+================+
//  +0   |     Magic      |    Version     |  Control Byte  |    Checksum    |
//       +----------------+----------------+----------------+----------------+
//  +4   |                        Total Packet Length                        |
//       +-------------------------------------------------------------------+
//  +8   |                           Payload Offset                          |
//       +-------------------------------------------------------------------+
//  +12  |                            User Header                            |
//  ...  |                                ...                                |
//       +-------------------------------------------------------------------+
//  +N   |                              Payload                              |
//  ...  |                                ...                                |
//       +-------------------------------------------------------------------+

Checksums

Each packet has a checksum, stored in the main header. Presently, checksums are calculated by adding every byte in the packet (including payload and user-defined header bytes, and excluding the checksum byte itself) and representing the result as a u8. Integer overflow is allowed while computing the checksum.

Checksums are verified by subtracting every byte (excluding the checksum byte) from the checksum stored in the header. If the result is 0, the sum is verified. Otherwise, it's not; the packet has been tampered with.

Fields
header: PacketHeader
Implementations

impl Packet

pub fn new() -> Packet

Constructor: creates an empty packet with default header information

pub fn to_bytes(&self) -> Vec<u8>

Returns an array of bytes that holds every byte in the packet

pub fn generate_checksum(&mut self)

Iterates through every byte in the packet (except for the checksum field) and performs a summation. This summation, represented by a u8, is placed in the checksum field of the packet's header: PacketHeader field.

pub fn verify_checksum(&self) -> bool

Iterates through every byte in the packet (except for the checksum field) and performs a summation. This summation, represented

pub fn user_header_append_byte(&mut self, byte: u8) -> bool

Attempts to append a single byte to the packet's user header. Returns true on success and false on failure. (Safe wrapper for self.user_header.append_byte())

pub fn user_header_append_bytes(&mut self, bytes: &mut Vec<u8>) -> bool

Attempts to append a vector of bytes to the packet's user header. Returns true on success and false on failure. (Safe wrapper for self.user_header.append_bytes())

pub fn user_header_append_string(&mut self, text: String) -> bool

Attempts to append a string to the packet's user header. Returns true on success and false on failure. (Safe wrapper for self.user_header.append_string())

pub fn payload_append_byte(&mut self, byte: u8) -> bool

Attempts to append a single byte to the packet's payload. Returns true on success and false on failure. (Safe wrapper for self.payload.append_byte())

pub fn payload_append_bytes(&mut self, bytes: &mut Vec<u8>) -> bool

Attempts to append a vector of bytes to the packet's payload. Returns true on success and false on failure. (Safe wrapper for self.payload.append_bytes())

pub fn payload_append_string(&mut self, text: String) -> bool

Attempts to append a string to the packet's payload. Returns true on success and false on failure. (Safe wrapper for self.payload.append_string())

pub fn get_user_header_data(&self) -> Vec<u8>

Returns a copy of the packet's user header data (if any)

pub fn get_payload_data(&self) -> Vec<u8>

pub fn to_string(&self) -> String

Generates a string representation of the packet and returns it