Module ssmp::messages::utils

Constants

HEADER_SZ

WRAP_SZ

Functions

decrypt_packet_if_set

encrypt_packet
creates a packet head for packet and then ecrypts packet

get_packet_sz_and_nonce
grab the packet encrypted size, or will return 0 if the packet was not encrypted

no_encrypt_packet
generate a packet head without encrypting the msg (zero set wrap)